Backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pyodbc==5.0.1
PyJWT==2.8.0
python-dotenv==1.0.0